feat(compiler): Enable format check and remove warnings
Check format-string args, and disable the warning of a few unused
functions, including signed/unsigned comparison warn.

+++ mbbsd/verifydb.cc
@@ -25,7 +25,7 @@ bool verifydb_transact(const void *msg, size_t len, Bytes *out) {
   rep.header.cb = sizeof(rep.header) + len;
   rep.header.operation = VERIFYDB_MESSAGE;
   if (towrite(fd, &rep.header, sizeof(rep.header)) != sizeof(rep.header) ||
-      towrite(fd, msg, len) != len)
+      towrite(fd, msg, len) != (int)len)
     return false;
 
   // Read reply.
@@ -33,7 +33,7 @@ bool verifydb_transact(const void *msg, size_t len, Bytes *out) {
     return false;
 
   out->resize(rep.header.cb - sizeof(rep.header));
-  if (toread(fd, &out->front(), out->size()) != out->size())
+  if (toread(fd, &out->front(), out->size()) != (int)out->size())
     return false;
 
   return true;
